requirements.txt
fastapi
uvicorn
pymongo
qdrant-client
llama-index
sentence-transformers
streamlit
facebook-sdk
python-dotenv